+++ src/raytracer/lights/directional/Directional.cpp
@@ -24,7 +24,7 @@ void raytracer::Directional::parse(unused const raytracer::Raytracer& raytracer,
     raytracer::Raytracer::setCFrame(descriptor, node);
 
     // Other settings
-    float intensity = 1.0f;
+    double intensity = 1.0f;
     if (node.lookupValue("intensity", intensity))
         this->_intensity = intensity;
